survai-platform/services/demo-bot/requirements.txt
# LiveKit Agents SDK
livekit-agents==1.4.2

# STT Provider
livekit-plugins-deepgram==1.4.2

# LLM Provider
livekit-plugins-openai==1.4.2

# TTS Provider
livekit-plugins-elevenlabs==1.4.2

# VAD (Voice Activity Detection)
livekit-plugins-silero==1.4.2

# Turn detection
livekit-plugins-turn-detector==1.4.2

# Noise cancellation (BVC — reduces VAD false triggers on phone calls)
livekit-plugins-noise-cancellation>=0.2.3

# Environment variables
python-dotenv==1.2.1

# Async HTTP
aiohttp==3.13.3
